third_party/WebKit/LayoutTests/shadow-dom/crashes/scrollbar-pseudo.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<style>
  #t1::before {
    display: block;
    width: 100px;
    height: 100px;
    overflow: scroll;
    content: ".";
  }
</style>
<div id="host"></div>
<div id="t1"></div>
<script>
  test(() => {
    host.attachShadow({mode: 'closed'});
    assert_true(true, "This test should not crash.");
  }, "Attaching v1 shadow should not cause DCHECK failure matching ::slotted rules for scrollbar pseudo.");
</script>
